--- spec/support/api.rb
@@ -1,7 +1,0 @@
-def api_prefix
-  "/api/#{Gitlab::API::VERSION}"
-end
-
-def json_response
-  JSON.parse(response.body)
-end
