+++ spec/support/api_helpers.rb
@@ -1,0 +1,9 @@
+module ApiHelpers
+  def api_prefix
+    "/api/#{Gitlab::API::VERSION}"
+  end
+
+  def json_response
+    JSON.parse(response.body)
+  end
+end
